Feature: rewind for all object almost ready
Basic feature works, all the objects in the TimeManagerList will rewind the them position when the player will rewind in the past, but still a bug on the rewind of the player. Some step are not found in the dictionary of position even if the step is register in it.